Move model to separate file
