+++ podcasts/HorizontalCollectionList.swift
@@ -1,22 +1,6 @@
 import SwiftUI
 import Foundation
 
-class HorizontalCollectionModel: ObservableObject {
-
-    @Published var colors: [Color] = [.blue, .green, .yellow, .orange, .pink, .purple, .cyan, .brown, .indigo]
-
-    var list: [[Color?]] {
-        return colors.pairs()
-    }
-}
-
-extension Color: @retroactive Identifiable {
-
-    public var id: String {
-        return description
-    }
-}
-
 struct HorizontalCollectionList: View {
 
     @StateObject var model: HorizontalCollectionModel = HorizontalCollectionModel()
